Added Comment and CommentAdapter, will be working on this to allow for opening moodHistory with comments
