+++ code/app/src/main/java/com/hamidat/nullpointersapp/models/CommentAdapter.java
@@ -1,0 +1,4 @@
+package com.hamidat.nullpointersapp.models;
+
+public class CommentAdapter {
+}
